Reorganized components, pushed uploads folder
